Moved MarkupItReactRenderer to /components. Added a story for it.
