+++ src/components/stories/index.js
@@ -2,3 +2,4 @@ import './Card';
 import './Icon';
 import './Toast';
 import './FindBar';
+import './MarkupItReactRenderer';
